+++ api/Domain/Entities/subject/Models/subject.cs
@@ -9,7 +9,7 @@ public partial class Subject : IEntity
   public required Class Class { get; set; }
   public required int Period { get; set; } 
   public bool IsActive { get; set; } = true;
-  public required float DurationHours { get; set; } 
+  public required double DurationHours { get; set; } 
   public DateOnly BeganAt { get; set; } 
   public ICollection<SpecificObjectives> SpecificObjectives { get; set; } = [];
   public ICollection<Post> Posts { get; set; } = [];
